trustifai-main/requirements.cpu.txt
# === core runtime deps (CPU-safe) ===
# --- core NLP ---
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
PyMuPDF

# --- langchain ecosystem ---
langchain
langchain-community
langchain-core
langchain-openai
langchainhub

# --- hf/torch stack (CPU) ---
torch          # keep unpinned; works on CPU. If needed: pip install --index-url https://download.pytorch.org/whl/cpu torch
transformers
accelerate
sentencepiece
peft

# --- data/graph/util ---
neo4j
chromadb
pandas
numpy
scikit-learn
networkx

# --- pdf & tooling ---
pypdf
pdfminer.six

# --- cli/config ---
typer
rich
python-dotenv
configparser


typing_extensions
rank_bm25
textwrap

langchain_huggingface
langchain_chroma
langchain_core